Prevents cursor being placed at start of file
Fixes #4 by checking if this is the first selection to be wrapped and if it is moving the cursor from the start of the document to the beginning of the buffer. All additional selections should receive new cursors at the beginning of their buffers. The first one has to be moved since selection.destroy() seems to place the cursor at the beginning of the file.